requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
distlib==0.3.4
fastapi==0.73.0
filelock==3.4.2
gunicorn==20.1.0
h11==0.13.0
identify==2.4.8
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
MarkupSafe==2.0.1
nodeenv==1.6.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.17.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pytesseract==0.3.9
pytest==7.0.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
virtualenv==20.13.1
